+++ dash/_utils.py
@@ -1,4 +1,6 @@
 import uuid
+import collections
+import six
 
 
 def interpolate_str(template, **data):
@@ -29,18 +31,22 @@ def generate_hash():
 
 def get_asset_path(
         requests_pathname,
-        routes_pathname,
         asset_path,
         asset_url_path):
-    i = requests_pathname.rfind(routes_pathname)
-    req = requests_pathname[:i]
 
     return '/'.join([
         # Only take the first part of the pathname
-        req,
+        requests_pathname.rstrip('/'),
         asset_url_path,
         asset_path
     ])
+
+
+# pylint: disable=no-member
+def patch_collections_abc(member):
+    if six.PY2:
+        return getattr(collections, member)
+    return getattr(collections.abc, member)
 
 
 class AttributeDict(dict):