+++ dash/dash.py
@@ -16,15 +16,16 @@ import logging
 
 from functools import wraps
 
-import plotly
-import dash_renderer
 import flask
 from flask import Flask, Response
 from flask_compress import Compress
 
-from .dependencies import Event, Input, Output, State
+import plotly
+import dash_renderer
+
+from .dependencies import Input, Output, State
 from .resources import Scripts, Css
-from .development.base_component import Component
+from .development.base_component import Component, ComponentRegistry
 from .development.validator import (DashValidator,
                                     generate_validation_error_message)
 from . import exceptions
@@ -32,8 +33,9 @@ from ._utils import AttributeDict as _AttributeDict
 from ._utils import interpolate_str as _interpolate
 from ._utils import format_tag as _format_tag
 from ._utils import generate_hash as _generate_hash
-from . import _watch
 from ._utils import get_asset_path as _get_asset_path
+from ._utils import patch_collections_abc as _patch_collections_abc
+from . import _watch
 from . import _configs
 
 
@@ -100,7 +102,7 @@ class Dash(object):
             name='__main__',
             server=None,
             static_folder='static',
-            assets_folder=None,
+            assets_folder='assets',
             assets_url_path='/assets',
             assets_ignore='',
             include_assets_files=True,
@@ -127,20 +129,14 @@ class Dash(object):
                 See https://github.com/plotly/dash/issues/141 for details.
                 ''', DeprecationWarning)
 
-        name = name if server is None else server.name
-        self._assets_folder = assets_folder or os.path.join(
-            flask.helpers.get_root_path(name), 'assets'
+        self._assets_folder = os.path.join(
+            flask.helpers.get_root_path(name),
+            assets_folder,
         )
         self._assets_url_path = assets_url_path
 
         # allow users to supply their own flask server
         self.server = server or Flask(name, static_folder=static_folder)
-
-        if 'assets' not in self.server.blueprints:
-            self.server.register_blueprint(
-                flask.Blueprint('assets', 'assets',
-                                static_folder=self._assets_folder,
-                                static_url_path=assets_url_path))
 
         env_configs = _configs.env_configs()
 
@@ -175,6 +171,22 @@ class Dash(object):
                 env_configs, 2678400))
         })
 
+        assets_blueprint_name = '{}{}'.format(
+            self.config.routes_pathname_prefix.replace('/', '_'),
+            'dash_assets'
+        )
+
+        self.server.register_blueprint(
+            flask.Blueprint(
+                assets_blueprint_name, name,
+                static_folder=self._assets_folder,
+                static_url_path='{}{}'.format(
+                    self.config.routes_pathname_prefix,
+                    assets_url_path.lstrip('/')
+                )
+            )
+        )
+
         # list of dependencies
         self.callback_map = {}
 
@@ -188,10 +200,9 @@ class Dash(object):
             Compress(self.server)
 
         @self.server.errorhandler(exceptions.PreventUpdate)
-        def _handle_error(error):
+        def _handle_error(_):
             """Handle a halted callback and return an empty 204 response"""
-            print(error, file=sys.stderr)
-            return ('', 204)
+            return '', 204
 
         # static files from the packages
         self.css = Css()
@@ -294,7 +305,7 @@ class Dash(object):
         return self._layout
 
     def _layout_value(self):
-        if isinstance(self._layout, collections.Callable):
+        if isinstance(self._layout, _patch_collections_abc('Callable')):
             self._cached_layout = self._layout()
         else:
             self._cached_layout = self._layout
@@ -303,7 +314,7 @@ class Dash(object):
     @layout.setter
     def layout(self, value):
         if (not isinstance(value, Component) and
-                not isinstance(value, collections.Callable)):
+                not isinstance(value, _patch_collections_abc('Callable'))):
             raise exceptions.NoLayoutException(
                 ''
                 'Layout must be a dash component '
@@ -420,11 +431,12 @@ class Dash(object):
                             namespace=resource['namespace']
                         ))
             elif 'external_url' in resource:
-                if isinstance(resource['external_url'], str):
-                    srcs.append(resource['external_url'])
-                else:
-                    for url in resource['external_url']:
-                        srcs.append(url)
+                if not is_dynamic_resource:
+                    if isinstance(resource['external_url'], str):
+                        srcs.append(resource['external_url'])
+                    else:
+                        for url in resource['external_url']:
+                            srcs.append(url)
             elif 'absolute_path' in resource:
                 raise Exception(
                     'Serving files from absolute_path isn\'t supported yet'
@@ -641,7 +653,6 @@ class Dash(object):
                 },
                 'inputs': v['inputs'],
                 'state': v['state'],
-                'events': v['events']
             } for k, v in self.callback_map.items()
         ])
 
@@ -652,9 +663,15 @@ class Dash(object):
             'Use `callback` instead. `callback` has a new syntax too, '
             'so make sure to call `help(app.callback)` to learn more.')
 
-    def _validate_callback_definition(self, output, inputs, state, events):
+    def _validate_callback(self, output, inputs, state):
         # pylint: disable=too-many-branches
         layout = self._cached_layout or self._layout_value()
+
+        for i in inputs:
+            if output == i:
+                raise exceptions.SameInputOutputException(
+                    'Same output and input: {}'.format(output)
+                )
 
         if (layout is None and
                 not self.config.first('suppress_callback_exceptions',
@@ -671,8 +688,7 @@ class Dash(object):
 
         for args, obj, name in [([output], Output, 'Output'),
                                 (inputs, Input, 'Input'),
-                                (state, State, 'State'),
-                                (events, Event, 'Event')]:
+                                (state, State, 'State')]:
 
             if not isinstance(args, list):
                 raise exceptions.IncorrectTypeException(
@@ -693,7 +709,7 @@ class Dash(object):
                                           'supress_callback_exceptions') and
                         arg.component_id not in layout and
                         arg.component_id != getattr(layout, 'id', None)):
-                    raise exceptions.NonExistantIdException('''
+                    raise exceptions.NonExistentIdException('''
                         Attempting to assign a callback to the
                         component with the id "{}" but no
                         components with id "{}" exist in the
@@ -726,7 +742,7 @@ class Dash(object):
                             component.available_properties and not
                             any(arg.component_property.startswith(w) for w in
                                 component.available_wildcard_properties)):
-                        raise exceptions.NonExistantPropException('''
+                        raise exceptions.NonExistentPropException('''
                             Attempting to assign a callback with
                             the property "{}" but the component
                             "{}" doesn't have "{}" as a property.\n
@@ -740,32 +756,20 @@ class Dash(object):
                             component.available_properties).replace(
                                 '    ', ''))
 
-                    if (hasattr(arg, 'component_event') and
-                            arg.component_event not in
-                            component.available_events):
-                        raise exceptions.NonExistantEventException('''
-                            Attempting to assign a callback with
-                            the event "{}" but the component
-                            "{}" doesn't have "{}" as an event.\n
-                            Here is a list of the available events in "{}":
-                            {}
-                        '''.format(
-                            arg.component_event,
-                            arg.component_id,
-                            arg.component_event,
-                            arg.component_id,
-                            component.available_events).replace('    ', ''))
+                    if hasattr(arg, 'component_event'):
+                        raise exceptions.NonExistentEventException('''
+                            Events have been removed.
+                            Use the associated property instead.
+                        ''')
 
-        if state and not events and not inputs:
-            raise exceptions.MissingEventsException('''
+        if state and not inputs:
+            raise exceptions.MissingInputsException('''
                 This callback has {} `State` {}
-                but no `Input` elements or `Event` elements.\n
-                Without `Input` or `Event` elements, this callback
+                but no `Input` elements.\n
+                Without `Input` elements, this callback
                 will never get called.\n
                 (Subscribing to input components will cause the
-                callback to be called whenever their values
-                change and subscribing to an event will cause the
-                callback to be called whenever the event is fired.)
+                callback to be called whenever their values change.)
             '''.format(
                 len(state),
                 'elements' if len(state) > 1 else 'element'
@@ -907,8 +911,8 @@ class Dash(object):
     # TODO - Check this map for recursive or other ill-defined non-tree
     # relationships
     # pylint: disable=dangerous-default-value
-    def callback(self, output, inputs=[], state=[], events=[]):
-        self._validate_callback_definition(output, inputs, state, events)
+    def callback(self, output, inputs=[], state=[]):
+        self._validate_callback(output, inputs, state)
 
         callback_id = '{}.{}'.format(
             output.component_id, output.component_property
@@ -921,10 +925,6 @@ class Dash(object):
             'state': [
                 {'id': c.component_id, 'property': c.component_property}
                 for c in state
-            ],
-            'events': [
-                {'id': c.component_id, 'event': c.component_event}
-                for c in events
             ]
         }
 
@@ -980,6 +980,21 @@ class Dash(object):
 
         target_id = '{}.{}'.format(output['id'], output['property'])
         args = []
+
+        flask.g.input_values = input_values = {
+            '{}.{}'.format(x['id'], x['property']): x.get('value')
+            for x in inputs
+        }
+        flask.g.state_values = {
+            '{}.{}'.format(x['id'], x['property']): x.get('value')
+            for x in state
+        }
+        changed_props = body.get('changedPropIds')
+        flask.g.triggered_inputs = [
+            {'prop_id': x, 'value': input_values[x]}
+            for x in changed_props
+        ] if changed_props else []
+
         for component_registration in self.callback_map[target_id]['inputs']:
             args.append([
                 c.get('value', None) for c in inputs if
@@ -1146,7 +1161,6 @@ class Dash(object):
     def get_asset_url(self, path):
         asset = _get_asset_path(
             self.config.requests_pathname_prefix,
-            self.config.routes_pathname_prefix,
             path,
             self._assets_url_path.lstrip('/')
         )
@@ -1253,9 +1267,20 @@ class Dash(object):
 
         if self._dev_tools.hot_reload:
             self._reload_hash = _generate_hash()
+
+            component_packages_dist = [
+                os.path.dirname(package.path)
+                if hasattr(package, 'path')
+                else package.filename
+                for package in (
+                    pkgutil.find_loader(x) for x in
+                    list(ComponentRegistry.registry) + ['dash_renderer']
+                )
+            ]
+
             self._watch_thread = threading.Thread(
                 target=lambda: _watch.watch(
-                    [self._assets_folder],
+                    [self._assets_folder] + component_packages_dist,
                     self._on_assets_change,
                     sleep_time=self._dev_tools.hot_reload_watch_interval)
             )
@@ -1274,42 +1299,43 @@ class Dash(object):
         self._hard_reload = True
         self._reload_hash = _generate_hash()
 
-        asset_path = os.path.relpath(
-            filename, os.path.commonprefix([self._assets_folder, filename]))\
-            .replace('\\', '/').lstrip('/')
+        if self._assets_folder in filename:
+            asset_path = os.path.relpath(
+                filename, os.path.commonprefix([self._assets_folder, filename])
+            ).replace('\\', '/').lstrip('/')
 
-        self._changed_assets.append({
-            'url': self.get_asset_url(asset_path),
-            'modified': int(modified),
-            'is_css': filename.endswith('css')
-        })
+            self._changed_assets.append({
+                'url': self.get_asset_url(asset_path),
+                'modified': int(modified),
+                'is_css': filename.endswith('css')
+            })
 
-        if filename not in self._assets_files and not deleted:
-            res = self._add_assets_resource(asset_path, filename)
-            if filename.endswith('js'):
-                self.scripts.append_script(res)
-            elif filename.endswith('css'):
-                self.css.append_css(res)
+            if filename not in self._assets_files and not deleted:
+                res = self._add_assets_resource(asset_path, filename)
+                if filename.endswith('js'):
+                    self.scripts.append_script(res)
+                elif filename.endswith('css'):
+                    self.css.append_css(res)
 
-        if deleted:
-            if filename in self._assets_files:
-                self._assets_files.remove(filename)
+            if deleted:
+                if filename in self._assets_files:
+                    self._assets_files.remove(filename)
 
-            def delete_resource(resources):
-                to_delete = None
-                for r in resources:
-                    if r.get('asset_path') == asset_path:
-                        to_delete = r
-                        break
-                if to_delete:
-                    resources.remove(to_delete)
+                def delete_resource(resources):
+                    to_delete = None
+                    for r in resources:
+                        if r.get('asset_path') == asset_path:
+                            to_delete = r
+                            break
+                    if to_delete:
+                        resources.remove(to_delete)
 
-            if filename.endswith('js'):
-                # pylint: disable=protected-access
-                delete_resource(self.scripts._resources._resources)
-            elif filename.endswith('css'):
-                # pylint: disable=protected-access
-                delete_resource(self.css._resources._resources)
+                if filename.endswith('js'):
+                    # pylint: disable=protected-access
+                    delete_resource(self.scripts._resources._resources)
+                elif filename.endswith('css'):
+                    # pylint: disable=protected-access
+                    delete_resource(self.css._resources._resources)
 
         self._lock.release()