+++ dash/exceptions.py
@@ -10,15 +10,15 @@ class CallbackException(DashException):
     pass
 
 
-class NonExistantIdException(CallbackException):
+class NonExistentIdException(CallbackException):
     pass
 
 
-class NonExistantPropException(CallbackException):
+class NonExistentPropException(CallbackException):
     pass
 
 
-class NonExistantEventException(CallbackException):
+class NonExistentEventException(CallbackException):
     pass
 
 
@@ -30,7 +30,7 @@ class IncorrectTypeException(CallbackException):
     pass
 
 
-class MissingEventsException(CallbackException):
+class MissingInputsException(CallbackException):
     pass
 
 
@@ -83,4 +83,12 @@ class DependencyException(DashException):
 
 
 class ResourceException(DashException):
+    pass
+
+
+class SameInputOutputException(CallbackException):
+    pass
+
+
+class MissingCallbackContextException(CallbackException):
     pass