+++ dash/extract-meta.js
@@ -8,7 +8,7 @@ const componentPaths = process.argv.slice(3);
 const ignorePattern = new RegExp(process.argv[2]);
 
 const excludedDocProps = [
-    'setProps', 'id', 'className', 'style', 'dashEvents', 'fireEvent'
+    'setProps', 'id', 'className', 'style'
 ];
 
 if (!componentPaths.length) {
@@ -42,7 +42,7 @@ function writeError(msg, filePath) {
 }
 
 function checkWarn(name, value) {
-    if (value.length < 1 && !excludedDocProps.includes(name.split('.').pop())) {
+    if (!value || (value.length < 1 && !excludedDocProps.includes(name.split('.').pop()))) {
         process.stderr.write(`\nDescription for ${name} is missing!\n`)
     }
 }